move persistence into systems
