+++ Assets/PetalsOfHope/Scripts/Systems/Persistence.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fb921b889b6c12e45838fa652e16ea76
+guid: 38d9cc321fec87e49bd57c82843d2247
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
